requirements.txt
pycryptodome
pysimplegui==4.60.5